Add topic_id returned by storage to the TopicMetadata (#1538)
* Add topic_id returned by storage to the TopicMetadata

- Rationale.
 To be able to distinguish topics by unique topic ID rather than by
 topic name in the future.


* Map inner int64_t topic_id to the external uint16_t topic_id in sqlite3


* Change TopicMetadata::id type to uint16_t


* Use atomic for last_extern_topic_id_


---------

+++ rosbag2_storage_sqlite3/src/rosbag2_storage_sqlite3/sqlite_wrapper.cpp
@@ -172,7 +172,7 @@ SqliteStatement SqliteWrapper::prepare_statement(const std::string & query)
   return std::make_shared<SqliteStatementWrapper>(db_ptr, query);
 }
 
-size_t SqliteWrapper::get_last_insert_id()
+int64_t SqliteWrapper::get_last_insert_id()
 {
   return sqlite3_last_insert_rowid(db_ptr);
 }
